requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
django-filter==24.2
djangorestframework==3.15.0
django-cors-headers==4.4.0
drf-spectacular==0.27.2
